@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { FormControl, FormGroup } from '@angular/forms';
+import { UntypedFormControl, UntypedFormGroup } from '@angular/forms';
 import { Post } from '../models/Post';
 import { UploadAPIService } from '../services/upload-api.service';
 
@@ -16,10 +16,10 @@ export class SingFileComponent implements OnInit {
 
   public image: any = null; //list of files
 
-  restFrom = new FormGroup({
-    name: new FormControl(''),
-    email: new FormControl(''),
-    phone: new FormControl('')
+  restFrom = new UntypedFormGroup({
+    name: new UntypedFormControl(''),
+    email: new UntypedFormControl(''),
+    phone: new UntypedFormControl('')
   })
 
 
