@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
+import { FormBuilder, UntypedFormControl, UntypedFormGroup, Validators } from '@angular/forms';
 import { Post } from '../models/Post';
 import { UploadAPIService } from '../services/upload-api.service';
 
@@ -17,10 +17,10 @@ export class TestComponent implements OnInit {
 
 
   
-  restFrom = new FormGroup({
-    name: new FormControl(''),
-    email: new FormControl(''),
-    phone: new FormControl('')
+  restFrom = new UntypedFormGroup({
+    name: new UntypedFormControl(''),
+    email: new UntypedFormControl(''),
+    phone: new UntypedFormControl('')
   })
 
   constructor(private uploader: UploadAPIService) { }
